+++ src/js/components/buy.js
@@ -1,0 +1,9 @@
+import { SELECTOR } from "../constants/constant.js"
+import { $ } from "../util.js"
+
+export const getBuyInput = () => {
+  const value = $(SELECTOR.BUY_INPUT).value
+  $(SELECTOR.BUY_INPUT).value = ""
+
+  return value
+}
