+++ src/js/components/pocket.js
@@ -1,0 +1,6 @@
+import { getQuotient } from "../util.js"
+import { TICKET } from "../constants/constant.js"
+
+export const getCount = (price) => {
+  return getQuotient(price, TICKET.PRICE)
+}
